refactor : memberService 에서 구현하도록 리팩토링
